Test antfu + husky v9

+++ eslint.config.mjs
@@ -1,9 +1,7 @@
-// eslint.config.mjs
 import antfu from '@antfu/eslint-config'
 
 export default antfu(
   {
-    // App vs lib (tunes some rules)
     type: 'app',
 
     // ESLint 9+ way to ignore files (replaces .eslintignore)
@@ -14,18 +12,14 @@ export default antfu(
       'pnpm-lock.yaml',
     ],
 
-    // Respect .gitignore as well (true by default; explicit for clarity)
     gitignore: true,
 
-    // TypeScript support
     typescript: {
       tsconfigPath: 'tsconfig.json',
     },
 
-    // React project
     react: true,
 
-    // Non-JS/TS formatters (via eslint-plugin-format)
     formatters: {
       css: true,
       html: true,
@@ -33,19 +27,16 @@ export default antfu(
       markdown: 'prettier',
     },
 
-    // Stylistic rules = your JS/TS formatter
     stylistic: {
       indent: 2,
-      quotes: 'double',
+      quotes: 'single',
       semi: true,
     },
   },
-
-  // Extra project-specific overrides (optional)
   {
     rules: {
+      // your overrides here
       // 'no-console': 'warn',
-      // 'react/jsx-no-target-blank': 'off',
     },
   },
 )
